+++ R/system.R
@@ -142,3 +142,128 @@ ps_tty_size <- function() {
   )
   c(width = ret[1], height = ret[2])
 }
+
+#' List all processes that loaded a shared library
+#'
+#' @details
+#' ## Notes:
+#' This function currently only works on Windows.
+#'
+#' On Windows, a 32 bit R process can only list other 32 bit processes.
+#' Similarly, a 64 bit R process can only list other 64 bit processes.
+#' This is a limitation of the Windows API.
+#'
+#' Even though Windows file systems are (almost always) case
+#' insensitive, the matching of `paths`, `user` and also `filter`
+#' are case sensitive. This might change in the future.
+#'
+#' This function can be very slow on Windows, because it needs to
+#' enumerate all shared libraries of all processes in the system,
+#' unless the `filter` argument is set. Make sure you set `filter`
+#' if you can.
+#'
+#' If you want to look up multiple shared libraries, list all of them
+#' in `paths`, instead of calling `ps_shared_lib_users` for each
+#' individually.
+#'
+#' If you are after libraries loaded by R processes, you might want to
+#' set `filter` to `c("Rgui.exe", "Rterm.exe", "rsession.exe")` The
+#' last one is for RStudio.
+#'
+#' @param paths Character vector of paths of shared libraries to
+#' look up. They must be absolute paths. They don't need to exist.
+#' Forward slashes are converted to backward slashes on Windows, and
+#' the output will always have backward slashes in the paths.
+#' @param user Character scalar or `NULL`. If not `NULL`, then only
+#' the processes of this user are considered. It defaults to the
+#' current user.
+#' @param filter Character vector or `NULL`. If not NULL, then it is
+#' a vector of glob expressions, used to filter the process names.
+#' @return A data frame (tibble) with columns:
+#' * `path`: path to the shared library,
+#' * `pid`: process ID of the process,
+#' * `name`: name of the process,
+#' * `username`: username of process owner,
+#' * `ps_handle`: `ps_handle` object, that can be used to further
+#'   query and manipulate the process.
+#'
+#' @export
+#' @family shared library tools
+#' @examplesIf ps::ps_is_supported() && ! ps:::is_cran_check() && ps::ps_os_type()[["WINDOWS"]]
+#' dlls <- vapply(getLoadedDLLs(), "[[", character(1), "path")
+#' psdll <- dlls[["ps"]][[1]]
+#' r_procs <- c("Rgui.exe", "Rterm.exe", "rsession.exe")
+#' ps_shared_lib_users(psdll, filter = r_procs)
+
+ps_shared_lib_users <- function(paths, user = ps_username(),
+                                filter = NULL) {
+  os <- ps_os_type()
+  if (!os[["WINDOWS"]]) {
+    stop("`ps_shared_lib_users()` currently only works on Windows")
+  }
+  assert_character(paths)
+  if (!is.null(user)) assert_string(user)
+  if (!is.null(filter)) assert_character(filter)
+  if (os[["WINDOWS"]]) paths <- gsub("/", "\\", paths, fixed = TRUE)
+
+  pids <- ps_pids()
+  processes <- not_null(lapply(pids, function(p) {
+    tryCatch(ps_handle(p), error = function(e) NULL) }))
+
+  nm <- map_chr(processes, function(p)
+    fallback(ps_name(p), NA_character_))
+
+  if (!is.null(filter)) {
+    selected <- glob$test_any(filter, nm)
+    processes <- processes[selected]
+    nm <- nm[selected]
+  }
+
+  us <- map_chr(processes, function(p)
+    fallback(ps_username(p), NA_character_))
+
+  if (!is.null(user)) {
+    us2 <- short_username(us)
+    selected <- (!is.na(us) & us == user) | (!is.na(us2) & us2 == user)
+    processes <- processes[selected]
+    nm <- nm[selected]
+    us <- us[selected]
+  }
+
+  libs <- lapply(processes, function(p) {
+    tryCatch(ps_shared_libs(p)$path, error = function(e) character())
+  })
+
+  # TODO: handle case insensitive OS/FS
+
+  match <- lapply(libs, intersect, paths)
+  match_len <- map_int(match, length)
+  match_processes <- processes[match_len > 0]
+  match_username <- us[match_len > 0]
+  match_name <- nm[match_len > 0]
+  match_len <- match_len[match_len > 0]
+  match_pids <- map_int(match_processes, ps_pid)
+
+  d <- data.frame(
+    stringsAsFactors = FALSE,
+    path = unlist(match),
+    pid = rep(match_pids, match_len),
+    name = rep(match_name, match_len),
+    username = rep(match_username, match_len),
+    ps_handle = I(match_processes)
+  )
+
+  # The ones without name probably finished already.
+  d <- d[!is.na(d$name), , drop = FALSE]
+
+  requireNamespace("tibble", quietly = TRUE)
+  class(d) <- unique(c("tbl_df", "tbl", class(d)))
+  d
+}
+
+short_username <- function(x) {
+  xs <- strsplit(x, "\\", fixed = TRUE)
+  p1 <- map_chr(xs, "[", 1)
+  p2 <- map_chr(xs, "[", 2)
+  ifelse(!is.na(p2), p2, x)
+}