+++ src/ps.h
@@ -42,6 +42,7 @@ SEXP psll_interrupt(SEXP p, SEXP ctrlc, SEXP interrupt_path);
 SEXP psll_connections(SEXP p);
 SEXP psll_get_nice(SEXP p);
 SEXP psll_set_nice(SEXP p, SEXP value);
+SEXP psll_dlls(SEXP p);
 
 /* System API */